shuffle: service barrier added (fixed concurrency issue)
