+++ Worker/src/main/scala/Service_WorkerShuffle.scala
@@ -1,13 +1,13 @@
 import Worker.WORKER_STATE
-import WorkerState.{SORT_PARTITION_FINISH, SORT_PARTITION_START}
-import protobuf.connect.{Empty, sortPartitionStartToSortPartitionFinishWorkerGrpc}
+import WorkerState.{SHUFFLE_SERVICE, SHUFFLE_START}
+import protobuf.connect.{Empty, shuffleStartToShuffleFinishWorkerGrpc}
 
 import scala.concurrent.Future
 
-class Service_WorkerSort extends sortPartitionStartToSortPartitionFinishWorkerGrpc.sortPartitionStartToSortPartitionFinishWorker {
-    override def startSorting(request: Empty): Future[Empty] = {
-        assert(WORKER_STATE == SORT_PARTITION_START)
-        WORKER_STATE = SORT_PARTITION_FINISH
+class Service_WorkerShuffle extends shuffleStartToShuffleFinishWorkerGrpc.shuffleStartToShuffleFinishWorker {
+    override def startShuffling(request: Empty): Future[Empty] = {
+        assert(WORKER_STATE == SHUFFLE_START)
+        WORKER_STATE = SHUFFLE_SERVICE
 
         Future.successful {
             Empty()
