+++ Worker/src/main/scala/Service_WorkerSort.scala
@@ -1,13 +1,13 @@
 import Worker.WORKER_STATE
-import WorkerState.{SHUFFLE_FINISH, SHUFFLE_START}
-import protobuf.connect.{Empty, shuffleStartToShuffleFinishWorkerGrpc}
+import WorkerState.{SORT_PARTITION_FINISH, SORT_PARTITION_START}
+import protobuf.connect.{Empty, sortPartitionStartToSortPartitionFinishWorkerGrpc}
 
 import scala.concurrent.Future
 
-class Service_WorkerShuffle extends shuffleStartToShuffleFinishWorkerGrpc.shuffleStartToShuffleFinishWorker {
-    override def startShuffling(request: Empty): Future[Empty] = {
-        assert(WORKER_STATE == SHUFFLE_START)
-        WORKER_STATE = SHUFFLE_FINISH
+class Service_WorkerSort extends sortPartitionStartToSortPartitionFinishWorkerGrpc.sortPartitionStartToSortPartitionFinishWorker {
+    override def startSorting(request: Empty): Future[Empty] = {
+        assert(WORKER_STATE == SORT_PARTITION_START)
+        WORKER_STATE = SORT_PARTITION_FINISH
 
         Future.successful {
             Empty()
