+++ Worker/src/main/scala/WorkerState.scala
@@ -4,6 +4,6 @@ object WorkerState extends Enumeration {
     CONNECTION_START, CONNECTION_FINISH,
     SAMPLING_START, SAMPLING_SAMPLE, SAMPLING_FINISH,
     SORT_PARTITION_START, SORT_PARTITION_FINISH,
-    SHUFFLE_START, SHUFFLE_FINISH,
+    SHUFFLE_START, SHUFFLE_SERVICE, SHUFFLE_FINISH,
     MERGE_START, MERGE_FINISH = Value
 }
